PR comments fix
