+++ packages/date-time-utilities/src/dateFormatting/dateFormatting.types.ts
@@ -1,0 +1,25 @@
+export interface IDateGridStrings {
+  /**
+   * An array of strings for the full names of months.
+   * The array is 0-based, so months[0] should be the full name of January.
+   */
+  months: string[];
+
+  /**
+   * An array of strings for the short names of months.
+   * The array is 0-based, so shortMonths[0] should be the short name of January.
+   */
+  shortMonths: string[];
+
+  /**
+   * An array of strings for the full names of days of the week.
+   * The array is 0-based, so days[0] should be the full name of Sunday.
+   */
+  days: string[];
+
+  /**
+   * An array of strings for the initials of the days of the week.
+   * The array is 0-based, so days[0] should be the initial of Sunday.
+   */
+  shortDays: string[];
+}
