+++ packages/date-time-utilities/src/dateFormatting/formatDay.ts
@@ -1,0 +1,5 @@
+/**
+ * Format date to a day string representation
+ * @param date - input date to format
+ */
+export const formatDay = (date: Date) => date.getDate().toString();
