+++ packages/date-time-utilities/src/dateFormatting/formatYear.ts
@@ -1,0 +1,5 @@
+/**
+ * Format date to a year string representation
+ * @param date - input date to format
+ */
+export const formatYear = (date: Date) => date.getFullYear().toString();
