+++ packages/date-time-utilities/src/index.ts
@@ -1,5 +1,16 @@
-export * from './dateMath/DateMath';
-export * from './dateValues/DateValues';
-export * from './dateValues/TimeConstants';
-export * from './dateGrid/DateGrid';
-export * from './dateGrid/DateGrid.types';
+export * from './dateMath/dateMath';
+export * from './dateValues/dateValues';
+export * from './dateValues/timeConstants';
+export * from './dateFormatting/dateFormatting.types';
+export * from './dateFormatting/formatDay';
+export * from './dateFormatting/formatMonthDayYear';
+export * from './dateFormatting/formatMonthYear';
+export * from './dateFormatting/formatYear';
+export * from './dateGrid/dateGrid.types';
+export * from './dateGrid/findAvailableDate';
+export * from './dateGrid/getBoundedDateRange';
+export * from './dateGrid/getDateRangeTypeToUse';
+export * from './dateGrid/getDayGrid';
+export * from './dateGrid/isAfterMaxDate';
+export * from './dateGrid/isBeforeMinDate';
+export * from './dateGrid/isRestrictedDate';
