+++ packages/date-time/etc/date-time.api.md
@@ -385,8 +385,8 @@ export interface IWeeklyDayPickerStyles extends Partial<ICalendarDayGridStyles> 
 export const WeeklyDayPicker: React.FunctionComponent<IWeeklyDayPickerProps>;
 
 
-export * from "@fluentui/date-time-utilities/lib/dateMath/DateMath";
-export * from "@fluentui/date-time-utilities/lib/dateValues/DateValues";
+export * from "@fluentui/date-time-utilities/lib/dateMath/dateMath";
+export * from "@fluentui/date-time-utilities/lib/dateValues/dateValues";
 
 // (No @packageDocumentation comment for this package)
