+++ packages/date-time/src/utilities/dateValues/TimeConstants.ts
@@ -1,1 +1,1 @@
-export * from '@fluentui/date-time-utilities/lib/dateValues/TimeConstants';
+export * from '@fluentui/date-time-utilities/lib/dateValues/timeConstants';
