+++ packages/office-ui-fabric-react/src/utilities/dateValues/DateValues.ts
@@ -1,1 +1,1 @@
-export * from '@fluentui/date-time-utilities/lib/dateValues/DateValues';
+export * from '@fluentui/date-time-utilities/lib/dateValues/dateValues';
